requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozendict==2.3.8
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.8
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
multitasking==0.0.11
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
TA-Lib==0.4.26
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.65.0
tweepy==4.14.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.3.2
wrapt==1.14.1
yfinance==0.2.18
zipp==3.15.0